Handled setting mob health and mobs RECEIVING damage.

+++ src/com/herocraftonline/dev/heroes/util/Properties.java
@@ -56,7 +56,7 @@ public class Properties {
     public boolean damageSystem;
     // Damage//
     public HashMap<Material, Integer> damageValues = new HashMap<Material, Integer>();
-    public HashMap<CreatureType, Double> mobMaxHealth = new HashMap<CreatureType, Double>();
+    public HashMap<CreatureType, Integer> mobMaxHealth = new HashMap<CreatureType, Integer>();
     public HashMap<CreatureType, Double> mobDamageValues = new HashMap<CreatureType, Double>();
     public HashMap<DamageCause, Integer> enviromentalDamageValues = new HashMap<DamageCause, Integer>();
